+++ src/App.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import TopPageContainer from 'container/TopPageContainer';
+import TopPageContainer from 'containers/TopPageContainer';
 
 const App: React.FC = () => {
   return (
